==> r/SouthWest/02_habitat_model/05_habitatfigures.R <==
###
# Project: ** GC Ch2 Habitat Spatial Predictions**
# Data:    BRUVS, BOSS Habitat data
# Task:    Habitat figures
# author:  Kingsley Griffin & Claude Spencer & Gabrielle Cummins
# date:    ** 6/09/2022 **
##

# This script makes report ready figures and plots

# CONTENTS
# Figure 1. Dominant habitat class map
# Figure 2. Individual habitat class predictions - facetted plot
# Figure 3. Bathymetry derivatives
# Figure 4. Predicted Relief
# Figure 5. Relief spatial random effect
# Figure 6. National Reef Model predictions

# Clear your environment
rm(list = ls())

# Load libraries
library(reshape2)
library(ggplot2)
library(viridis)
library(terra)
library(patchwork)
library(sf)
library(sfheaders)
library(dplyr)
library(rgdal)
library(stars)
library(smoothr)
library(ggnewscale)
library(gridExtra)

# Set your study name
name <- "SouthWest"                                                              # Change here

# Set CRS for shapefiles
wgscrs <- "+proj=longlat +datum=WGS84"                                    # Lat long projection
     
# Bring in spatial layers
# Load aus outline, state and commonwealth marine parks
aus     <- st_read("data/spatial/shapefiles/cstauscd_r.mif")                    # Geodata 100k coastline available: https://data.gov.au/dataset/ds-ga-a05f7892-eae3-7506-e044-00144fdd4fa6/
aus     <- aus[aus$FEAT_CODE == "mainland", ]                                   # Add islands here if needed
aumpa   <- st_read("data/spatial/shapefiles/AustraliaNetworkMarineParks.shp")   # All aus mpas
st_crs(aus) <- st_crs(aumpa)                                                    # Set CRS to match - WGS84 and GDA94 effectively the same
aus    <- aus[!aus$FEAT_CODE == c("sea"), ]
e <- ext(114, 116, -35, -33)                                                 # Change your extent here
mpa <- st_crop(aumpa, e)                                                        # All commonwealth zones in the study area
npz <- mpa[mpa$ZoneName %in% "National Park Zone", ]                            # Only National Park Zones in the study area

# Load Jac's National Reef Model
# jacmap  <- raster("data/spatial/rasters/ecosystem-types-19class-naland.tif")    # Jac's aus habitat map
# jacmap  <- crop(jacmap, e)                                                      # Crop to the general study area

# Load tidy habitat data from 03_mergedata
habi    <- readRDS(paste(paste0('data/tidy/', name), 
                         'habitat-bathy-derivatives.rds', sep = "_")) %>%
  glimpse()

# Load terrestrial parks
# terrnp <- st_read("data/spatial/shp/Legislated_Lands_and_Waters_DBCA_011.shp") %>%  # Terrestrial reserves
#   dplyr::filter(leg_catego %in% c("Nature Reserve", "National Park"))
# terrnp <- st_crop(terrnp, xmin = 113, ymin = -30, xmax = 116, ymax = -26)       # Just abrolhos area

# Load the coastal waters boundary
cwatr  <- st_read('data/spatial/shapefiles/amb_coastal_waters_limit.shp')              

# Load bathymetry data
bathdf <- readRDS(paste(paste0('data/spatial/rasters/',                         
                               name), 'ga_bathy.rds', sep = "_"))

# read in spatial predictions from 'R/05_habitat_model.R'
spreddf <- readRDS(paste(paste0('output/SWC/', name),
                         'spatial_habitat_predictions.rds', sep = "_")) %>%
  dplyr::mutate(dom_tag = as.factor(dom_tag)) %>%                               # Factorise
  dplyr::mutate(dom_tag = dplyr::recode(dom_tag,                                # Tidy names for plot legend
                                 seagrasses.fit = "Seagrasses",
                                 # reef = "Reef",
                                 macroalg.fit = "Macroalgae",
                                 rock.fit = "Rock",
                                 sand.fit = "Sand",
                                 inverts.fit = "Sessile invertebrates")) %>%
  glimpse()

# Figure 1: Categorical habitat maps ----
# Assign habitat class colours
hab_cols <- scale_fill_manual(values = c("Seagrasses" = "#E69F00",
                                         # "Reef" = "#c8cbcf",
                                         "Macroalgae" = "seagreen",
                                         "Rock" = "#D55E00",
                                         "Sand" = "#F0E442",
                                         "Sessile invertebrates" = "#56B4E9"
))

npz_cols <- scale_colour_manual(values = c("National Park Zone" = "#7bbc63"),
                                name = "Australian Marine Parks")

#Build plot elements for site 1
p1 <- ggplot() +
  geom_tile(data = spreddf, aes(x, y, fill = dom_tag)) +
  hab_cols +                                                                    # Class colours
  geom_contour(data = bathdf, aes(x = x, y = y, z = Z),                         # Contour lines
               breaks = c(- 30, -70, - 200),                                 # Contour breaks - change to binwidth for regular contours
               colour = "#000000",
               alpha = 1, size = 0.5) +
  geom_sf(data = npz, fill = NA, aes(colour = ZoneName), linewidth = 0.5) +                          # Add national park zones
  npz_cols+
  #new_scale_colour() +
  geom_sf(data = aus, fill = "seashell2", colour = "grey80", size = 0.5) +       #trying to add in AUSMAP
  coord_sf(xlim = c(114.2, 115.2),                              # Set plot limits
          ylim = c(-34.2, -33.5)) +
  labs(x = NULL, y = NULL, fill = "Habitat",                                    # Labels  
       colour = NULL, title = "Capes Region, South West") +
     annotate("text", x = c(115.0, 114.79, 114.51),          # Add contour labels manually
           y = c(-34.0, -34.0, -34.0), 
           label = c("30m", "70m", "200m"),
           size = 2, colour = "#000000") +
  theme_minimal()+
  theme(
    plot.title = element_text(size = 14, hjust = 0.5)  # Center title horizontally
  )
png(filename = paste(paste("plots", name, sep = "/"),                   # Save output
                     "dominant_habitat.png", sep = "_"),
    width = 8, height = 4, res = 300, units = "in")                             # Change the dimensions here as necessary
p1
dev.off()


# Figure 2. Individual habitat class predictions ----
# Melt classes for faceting
widehabitfit <- spreddf %>%
  dplyr:: select(pseagrasses.fit, pmacroalg.fit, psand.fit, prock.fit, pinverts.fit, x, y) %>%
  tidyr::pivot_longer(cols = starts_with("p"),                                  # Careful here that you don't have any other columns starting with 'p'
                      values_to = "value", names_to = "variable") %>%
  dplyr::mutate(variable = dplyr::recode(variable,                              # Tidy variable names
                pseagrasses.fit = "Seagrasses",
                pmacroalg.fit = "Macroalgae",
                prock.fit = "Rock",
                psand.fit = "Sand",
                pinverts.fit = "Sessile Invertebrates")) %>%
  glimpse()

###GC to make below SE dataframe
widehabitse <- spreddf %>%
  dplyr::select(pseagrasses.se.fit, pmacroalg.se.fit, psand.se.fit, prock.se.fit, pinverts.se.fit, x, y) %>%
  tidyr::pivot_longer(cols = starts_with("p"),                                  # Careful here that you don't have any other columns starting with 'p'
                      values_to = "value", names_to = "variable") %>%
  dplyr::mutate(variable = dplyr::recode(variable,                              # Tidy variable names
                                         pseagrasses.se.fit = "Seagrasses SE",
                                         pmacroalg.se.fit = "Macroalgae SE",
                                         prock.se.fit = "Rock SE",
                                         psand.se.fit = "Sand SE",
                                         pinverts.se.fit = "Sessile Invertebrates SE")) %>%
  glimpse()

# Make a dataframe for your contour line annotations - doesn't work otherwise for facetted plots
dep_ann <- data.frame(x = c(115.02, 114.8, 114.55),                            
                      y = c(-34.0, -34.0, -34.0),
                      label = c("30m", "70m", "200m"))

# Build the plot for predicted Habitat individual classes
p22 <- ggplot() + 
  geom_tile(data = widehabitfit, 
            aes(x, y, fill = value)) +
  scale_fill_viridis(direction = -1, limits = c(0, max(widehabitfit$value))) +
  guides(fill = guide_colourbar(title.position = "top", title.hjust = 0.5))+
  new_scale_colour()+
  geom_sf(data = aus, fill = "seashell2", colour = "grey80", size = 0.5) +
  geom_contour(data = bathdf, aes(x, y, z = Z),                                 # Contour lines
               breaks = c(-30, -70, -200), colour = "#000000",
               alpha = 1, size = 0.5) +
  geom_text(data = dep_ann,aes(x,y,label = label),
            inherit.aes = F, size = 2, colour = "#000000") +
  new_scale_colour()+
    geom_sf(data = npz, fill = NA, aes(colour = ZoneName), linewidth = 0.5, show.legend = FALSE) +                          # National park zones
  npz_cols+
  coord_sf(xlim = c(114.4, 115.05),                              # Set plot limits
           ylim = c(-34.2, -33.5)) +
  labs(x = NULL, y = NULL, fill = "Habitat (mean)") +      # Labels
  theme_minimal(base_size = 9) +
  theme(legend.position = "bottom",
        legend.box.margin = margin(1,1,1,90),
        legend.key.width = unit(1, "cm"),
        legend.text = element_text(size =15),
        legend.title = element_text(size =15),
        plot.margin = margin(1,1,1,1), 
        axis.text.y = element_text(size =15),
    axis.text.x = element_text(angle = 90, size = 15),
    strip.text = element_text(size = 11, margin = margin())) +
  facet_wrap(~variable, ncol = 1)                                               # Facet for each variable

png(filename = paste(paste("plots", name, sep = "/"),                   # Save the output
                     "habitat_class_predicted.png", sep = "_"),
    width = 12, height = 15, res = 400, units = "in")                             # Change the dimensions here as necessary
print(p22)
p22
dev.off()


## SE FIGURE###
p23 <- ggplot() + 
  geom_tile(data = widehabitse,
            aes(x, y, fill = value)) +
  scale_fill_viridis(direction = -1, option = "C", limits = c(0.00, 0.12), breaks=seq(0.00,0.12,by=0.03)) +
  guides(fill = guide_colourbar(title.position = "top", title.hjust = 0.5))+
  new_scale_colour()+
  geom_sf(data = aus, fill = "seashell2", colour = "grey80", size = 0.5) +
  geom_contour(data = bathdf, aes(x, y, z = Z),                                 # Contour lines
               breaks = c(-30, -70, -200), colour = "#000000",
               alpha = 1, size = 0.5) +
  geom_text(data = dep_ann,aes(x,y,label = label),
            inherit.aes = F, size = 2, colour = "#000000") +
  new_scale_colour()+
  geom_sf(data = npz, fill = NA, aes(colour = ZoneName), linewidth = 0.5) +
  npz_cols+
  guides(colour = guide_legend(title.position = "top", title.hjust = 0.5))+
  coord_sf(xlim = c(114.4, 115.05),                              # Set plot limits
           ylim = c(-34.2, -33.5)) +
  labs(x = NULL, y = NULL, fill = "Habitat (SE)") +      # Labels
  theme_minimal(base_size = 9) +
  theme(legend.position = "bottom",
        legend.box.margin=margin(10,240,10,10),
        legend.key.width = unit(1, "cm"),
        legend.text = element_text(size =15),
        legend.title = element_text(size =15),
        plot.margin = margin(1,1,1,1), 
        axis.text.y = element_text(size =15),
    axis.text.x = element_text(angle = 90, size = 15),
    strip.text = element_text(size = 11, margin = margin()))+
  facet_wrap(~variable, ncol = 1)                                               # Facet for each variable
  
 
    # geom_sf(data = aus, fill = "seashell2", colour = "grey80", size = 0.5) +
  
    
  png(filename = paste(paste("plots", name, sep = "/"),                   # Save the output
                       "habitat_SE_predicted.png", sep = "_"),
      width =12, height = 15, res = 400, units = "in") 
  print(p23)
  p23 
  dev.off()

  
  # p24 <- grid.arrange(
  #   p22, p23, top = grid::textGrob('Capes Region, South West', gp=grid::gpar(fontsize=24)),
  #   nrow=1, ncol=2)
  # 
  # # png(filename = paste(paste("plots", name, sep = "/"),                   # Save the output
  # #                      "indi_combined_predicted.png", sep = "_"),
  # #     width = 20, height = 15, res = 400, units = "in")  
  # 
  # print(p24)
  # dev.off()
  # 
  
  #Combine p22 and p23
  p24 <- p22 + p23

  # Add a title to the combined plot
  combined_plot <- p24 +
    ggtitle("Capes Region, South West") +
    theme(plot.title = element_text(size = 18, hjust = 0.5))  # Adjust the size as needed

  # Save or display the combined plot
  png(
    filename = paste(paste("plots", name, sep = "/"),
                     "combined_habitat_plots.png", sep = "_"),
    width = 12, height = 15, res = 400, units = "in"
  )
  print(combined_plot)
  dev.off()
  
  # # Combine p22 and p23 using grid.arrange and add a common title
  # combined_plot <- grid.arrange(
  #   p22, p23, nrow = 1, top = "Capes Region, South West"
  # )
  # 
  # # Save or display the combined plot
  # png(
  #   filename = paste(paste("plots", name, sep = "/"), 
  #                    "combined_habitat_plots.png", sep = "_"),
  #   width = 15, height = 15, res = 400, units = "in"
  # )
  # print(combined_plot)
  # dev.off()
  
  
# # Figure 3. Bathymetry derivatives ----
# # depth
# pd <- ggplot() +
#   geom_tile(data = spreddf[spreddf$sitens == 1, ], aes(x, y, fill = depth)) +
#   scale_fill_viridis(option = "A", direction = -1,
#                      limits = c(0, max(spreddf$depth))) +
#   geom_sf(data = ab_npz[ab_npz$parkid == 3, ], fill = NA, colour = "#7bbc63") +
#   labs(x= NULL, y = NULL, title = "Big Bank") +
#   guides(fill = "none") +
#   theme_minimal()
# 
# pdb <- ggplot() +
#   geom_tile(data = spreddf[spreddf$sitens == 0, ], aes(x, y, fill = depth)) +
#   scale_fill_viridis(option = "A", direction = -1,
#                      limits = c(0, max(spreddf$depth))) +
#   geom_sf(data = ab_npz[ab_npz$parkid == 2, ], fill = NA, colour = "#7bbc63") +
#   labs(x= NULL, y = NULL,
#        fill = "Depth", title = "Shallow Bank") +
#   theme_minimal()
# pd + pdb
# 
# # tpi
# pt <- ggplot() +
#   geom_tile(data = spreddf[spreddf$sitens == 1, ], aes(x, y, fill = tpi)) +
#   scale_fill_viridis(option = "D", direction = 1,
#                      limits = c( min(spreddf$tpi), max(spreddf$tpi))) +
#   geom_sf(data = ab_npz[ab_npz$parkid == 3, ], fill = NA, colour = "#7bbc63") +
#   labs(x= NULL, y = NULL) +
#   guides(fill = "none") +
#   theme_minimal()
# ptb <- ggplot() +
#   geom_tile(data = spreddf[spreddf$sitens == 0, ], aes(x, y, fill = tpi)) +
#   scale_fill_viridis(option = "D", direction = 1,
#                      limits = c( min(spreddf$tpi), max(spreddf$tpi))) +
#   geom_sf(data = ab_npz[ab_npz$parkid == 2, ], fill = NA, colour = "#7bbc63") +
#   labs(x= NULL, y = NULL,
#        fill = "TPI") +
#   theme_minimal()
# pt + ptb
# 
# # roughness
# pr <- ggplot() +
#   geom_tile(data = spreddf[spreddf$sitens == 1, ], aes(x, y, fill = roughness)) +
#   scale_fill_viridis(option = "D", direction = 1,
#                      limits = c( min(spreddf$roughness), max(spreddf$roughness))) +
#   geom_sf(data = ab_npz[ab_npz$parkid == 3, ], fill = NA, colour = "#7bbc63") +
#   labs(x= NULL, y = NULL) +
#   guides(fill = "none") +
#   theme_minimal() 
# prb <- ggplot() +
#   geom_tile(data = spreddf[spreddf$sitens == 0, ], aes(x, y, fill = roughness)) +
#   scale_fill_viridis(option = "D", direction = 1,
#                      limits = c(min(spreddf$roughness), max(spreddf$roughness))) +
#   geom_sf(data = ab_npz[ab_npz$parkid == 2, ], fill = NA, colour = "#7bbc63") +
#   labs(x= NULL, y = NULL,
#        fill = "Roughness") +
#   theme_minimal()
# pr + prb
# 
# # detrended
# pdt <- ggplot() +
#   geom_tile(data = spreddf[spreddf$sitens == 1, ], aes(x, y, fill = detrended)) +
#   scale_fill_viridis(option = "D", direction = 1,
#                      limits = c( min(spreddf$detrended), max(spreddf$detrended))) +
#   geom_sf(data = ab_npz[ab_npz$parkid == 3, ], fill = NA, colour = "#7bbc63") +
#   labs(x= NULL, y = NULL) +
#   guides(fill = "none") +
#   theme_minimal()
# pdtb <- ggplot() +
#   geom_tile(data = spreddf[spreddf$sitens == 0, ], aes(x, y, fill = detrended)) +
#   scale_fill_viridis(option = "D", direction = 1,
#                      limits = c(min(spreddf$detrended), max(spreddf$detrended))) +
#   geom_sf(data = ab_npz[ab_npz$parkid == 2, ], fill = NA, colour = "#7bbc63") +
#   labs(x= NULL, y = NULL,
#        fill = "Detrended") +
#   theme_minimal() 
# pdt + pdtb
# 
# # Figure 4. Predicted relief ----## relevant for fish things
# pcelldf <- readRDS('output/predicted_relief_site.rds')
# pcelldf$sitens <- ifelse(pcelldf$y > 6940000, 1, 0)
# pcelldf$prelief[pcelldf$prelief < 0] <- 0
# 
# p4 <- ggplot() +
#   geom_tile(data = pcelldf[pcelldf$sitens == 1, ], aes(x, y, fill = prelief)) +
#   scale_fill_viridis(option = "C", direction = -1, 
#                      limits = c(0, max(pcelldf$prelief))) +
#   geom_sf(data = ab_npz[ab_npz$parkid == 3, ], fill = NA, colour = "#7bbc63") +
#   labs(x= NULL, y = NULL, 
#        fill = "p. relief") +
#   guides(fill = "none") +
#   theme_minimal()
# 
# p42 <- ggplot() +
#   geom_tile(data = pcelldf[pcelldf$sitens == 0, ], aes(x, y, fill = prelief)) +
#   scale_fill_viridis(option = "C", direction = -1, 
#                      limits = c(0, max(pcelldf$prelief))) +
#   geom_sf(data = ab_npz[ab_npz$parkid == 2, ], fill = NA, colour = "#7bbc63") +
#   labs(x= NULL, y = NULL, 
#        fill = "Relief score") +
#   theme_minimal()
# 
# # relief only
# p4 + p42 + plot_layout(widths = c(0.44, 0.56))
# ggsave("plots/spatial/site_relief_p.png", width = 10, height = 6, dpi = 160)
# 
# # combined spatial layers
# 
# (pd + pdb)   /
#   (pt + ptb)   /
#   (pr + prb)   /
#   (pdt + pdtb) /
#   (p4 + p42)   +
#   plot_layout(widths = c(0.44, 0.56)) &
#   theme(text = element_text(size = 8))
# ggsave("plots/spatial/site_spatial_layers.png", width = 10, height = 12, dpi = 160)
# 
# 
# # Figure 5. Relief spatial random effect
# p5 <- ggplot() +
#   geom_tile(data = pcelldf[pcelldf$sitens == 1, ], aes(x, y, fill = p_sp)) +
#   scale_fill_viridis(option = "B", 
#                      limits = c(min(pcelldf$p_sp), max(pcelldf$p_sp))) +
#   geom_sf(data = ab_npz[ab_npz$parkid == 3, ], fill = NA, colour = "#7bbc63") +
#   geom_point(data = habi[habi$ns == 1, ], aes(longitude.1, latitude.1), 
#              alpha = 0.7, colour = "grey70", size = 1, shape = 3) +
#   labs(x= NULL, y = NULL, title = "Big Bank") +
#   guides(fill = "none") +
#   theme_minimal()
# 
# p52 <- ggplot() +
#   geom_tile(data = pcelldf[pcelldf$sitens == 0, ], aes(x, y, fill = p_sp)) +
#   scale_fill_viridis(option = "B", 
#                      limits = c(min(pcelldf$p_sp), max(pcelldf$p_sp))) +
#   geom_sf(data = ab_npz[ab_npz$parkid == 2, ], fill = NA, colour = "#7bbc63") +
#   geom_point(data = habi[habi$ns == 0, ], aes(longitude.1, latitude.1), 
#              alpha = 0.7, colour = "grey70", size = 2, shape = 3) +
#   labs(x= NULL, y = NULL, 
#        fill = "Spatial\ndependence", title = "Shallow Bank") +
#   theme_minimal()
# 
# p5 + p52 + plot_layout(widths = c(0.44, 0.56))
# ggsave("plots/spatial/site_relief_spatialeffect.png", 
#        width = 10, height = 3, dpi = 160)
# 
# # Figure 6. National Reef Model predictions
# # jac's map, eh
# # sort out the classes
# sitebathy <- readRDS('output/ga_bathy_fine.rds')                                # finer bathy
# colnames(sitebathy)[3] <- "Depth"
# # sitebathy <- sitebathy[sitebathy$Depth > -1000, ]                               # trim to reduce legend
# sitebathy <- sitebathy[sitebathy$x > 112.7 & sitebathy$x < 114.4, ] 
# sitebathy <- sitebathy[sitebathy$y > -28.4 & sitebathy$y < -26.6, ]
# 
# jlevs  <- ratify(jacmap)
# jclass <- levels(jlevs)[[1]]
# jclass[["class"]] <- c("Shelf.unvegetated.soft.sediments",
#                        "Upper.slope.unvegetated.soft.sediments", 
#                        "Mid.slope.sediments",
#                        "Lower.slope.reef.and.sediments",
#                        "Abyssal.reef.and.sediments", 
#                        "Seamount.soft.sediments", 
#                        "Shelf.vegetated.sediments", 
#                        "Shallow.coral.reefs.less.than.30.m.depth", 
#                        "Mesophotic.coral.reefs", 
#                        "Rariophotic.shelf.reefs", 
#                        "Upper.slope.rocky.reefs.shelf.break.to.700.m.depth", 
#                        "Mid.slope.reef", 
#                        "Artificial.reefs.pipelines.and.cables")                 # The class names
# levels(jacmap) <- jclass
# 
# jmap_df <- as.data.frame(jacmap, xy = TRUE, na.rm = TRUE)
# colnames(jmap_df)[3] <- "classname"
# jmap_df$classname <- gsub("\\.", " ", jmap_df$classname)                        # Replace . with space in names
# 
# jcls_cols <- scale_fill_manual(values = c(
#   "Shallow coral reefs less than 30 m depth" = "coral2", 
#   "Shelf unvegetated soft sediments" = "cornsilk1",
#   "Shelf vegetated sediments" = "seagreen3",
#   "Mesophotic coral reefs" = "darkorange3",
#   "Rariophotic shelf reefs" = "steelblue3",
#   "Upper slope unvegetated soft sediments" = "wheat1",
#   "Mid slope sediments" = "#f7d29c"))  # navajowhite1 - made slightly darker
# 
# waterr_cols <- scale_fill_manual(values = c("National Park" = "#c4cea6",
#                                             "Nature Reserve" = "#e4d0bb"),
#                                  guide = "none")
# 
# # assign mpa colours - full levels are saved at end of script for future ref
# nmpa_cols <- scale_color_manual(breaks = c("National Park Zone", 
#                                            "Multiple Use Zone", 
#                                            "Special Purpose Zone"), 
#                                 values = c("National Park Zone" = "#7bbc63",
#                                           "Multiple Use Zone" = "#b9e6fb",
#                                           "Special Purpose Zone" = "#6daff4"
# ))
# 
# ab_nmp$ZoneName <- factor(ab_nmp$ZoneName, levels = c("Multiple Use Zone", "Special Purpose Zone",
#                                                       "Habitat Protection Zone","National Park Zone"
#                                                       ))
# 
# 
# p7 <- ggplot() +
#   geom_sf(data = aus, fill = "seashell2", colour = "grey80", size = 0.1) +
#   geom_sf(data = terrnp, aes(fill = leg_catego), alpha = 4/5, colour = NA) +
#   waterr_cols +
#   new_scale_fill() +  
#   geom_tile(data = jmap_df, aes(x, y, fill = classname)) +
#   jcls_cols +
#   geom_contour(data = sitebathy, aes(x = x, y = y, z = Depth),
#                breaks = c(-30, -70, -200, - 700, - 7000), colour = "black", alpha = 1, size = 0.18) +
#   geom_sf(data = ab_nmp, fill = NA, aes(colour = ZoneName)) +
#   nmpa_cols+
#   labs(color = "Australian Marine Parks") +
#   new_scale_color() +
#   geom_sf(data = cwatr, colour = "firebrick", alpha = 4/5, size = 0.2) +
#   # annotate("rect", xmin = 113.02, xmax = 113.29, ymin = -27.19, ymax = -27.08,
#   #          colour = "grey25", fill = "white", alpha = 1/5, size = 0.2) +
#   # annotate("text", x = 113.15, y = -27.05, size = 3, 
#   #          colour = "grey20", label = "swabrnpz09") +
#   # annotate("rect", xmin = 113.24, xmax = 113.58, ymin = -28.13, ymax = -28.02,
#   #          colour = "grey25", fill = "white", alpha = 1/5, size = 0.2) +
#   # annotate("text", x = 113.42, y = -27.99, size = 3,
#   #          colour = "grey20", label = "swabrnpz06") +
#   annotate("text", y = c(-27.875, -27.875,-27.875,-27.875, -26.87), 
#            x = c(114.07,  113.32, 113.15, 112.79, 113.167), 
#            label = c("30m", "70m", "200m", "700m", "70m"), size = 2) +
#   coord_sf(xlim = c(112.8, 114.3), ylim = c(-28.25, -26.7)) +
#   labs(fill = "Habitat classification", x = NULL, y = NULL) +
#   annotate("point", y = c(-27.7115), x = c(114.1714), size = 0.75) +
#   annotate("text", y = c(-27.7115), x = c(114.275),
#            label = c("Kalbarri"), size = 3) +
#   theme_minimal()
# 
# png(filename = "plots/spatial/site_jmonk_natmap.png", width = 8, height = 6,
#     units = "in", res = 200)
# p7
# dev.off()
